--- src/content/app/entity-viewer/transcript-view/components/transcript-exons/exons-continuous/ExonsContinuous.module.css
@@ -1,55 +1,0 @@
-.container {
-  display: grid;
-  grid-template-rows: auto auto;
-  row-gap: 1.5rem;
-  padding-left: var(--double-standard-gutter);
-}
-
-.rowGrid {
-  display: grid;
-  grid-template-columns: min-content max-content min-content;
-  column-gap: 16px;
-}
-
-.row {
-  display: grid;
-  grid-template-columns: subgrid;
-  grid-column: 1 / -1;
-  align-items: baseline;
-  padding: 0.8rem 0;
-}
-
-.rowLeft, .rowRight {
-  font-size: 11px;
-  font-weight: var(--font-weight-light);
-  user-select: none;
-}
-
-.sequence {
-  font-family: var(--font-family-monospace);
-}
-
-.firstBase, .lastBase {
-  position: relative;
-}
-
-.featureName {
-  position: absolute;
-  top: 0;
-  left: 50%;
-  font-size: 11px;
-  font-weight: var(--font-weight-light);
-  transform: translateY(-100%);
-  white-space: nowrap;
-  user-select: none;
-}
-
-.boundaryMarker {
-  position: absolute;
-  top: -5px;
-  right: -100%;
-  transform: translateX(-30%);
-  width: 10px;
-  fill: var(--color-orange);
-  stroke: var(--color-orange);
-}
